+++ src/core/drd_server_runtime.h
@@ -15,6 +15,12 @@ G_BEGIN_DECLS
 #define DRD_TYPE_SERVER_RUNTIME (drd_server_runtime_get_type())
 G_DECLARE_FINAL_TYPE(DrdServerRuntime, drd_server_runtime, DRD, SERVER_RUNTIME, GObject)
 
+typedef enum
+{
+    DRD_FRAME_TRANSPORT_SURFACE_BITS = 0,
+    DRD_FRAME_TRANSPORT_GRAPHICS_PIPELINE
+} DrdFrameTransport;
+
 DrdServerRuntime *drd_server_runtime_new(void);
 
 DrdCaptureManager *drd_server_runtime_get_capture(DrdServerRuntime *self);
@@ -30,6 +36,8 @@ gboolean drd_server_runtime_pull_encoded_frame(DrdServerRuntime *self,
                                                 DrdEncodedFrame **out_frame,
                                                 GError **error);
 
+void drd_server_runtime_set_transport(DrdServerRuntime *self, DrdFrameTransport transport);
+DrdFrameTransport drd_server_runtime_get_transport(DrdServerRuntime *self);
 DrdFrameCodec drd_server_runtime_get_codec(DrdServerRuntime *self);
 gboolean drd_server_runtime_get_encoding_options(DrdServerRuntime *self,
                                                   DrdEncodingOptions *out_options);