+++ src/session/drd_rdp_session.h
@@ -3,6 +3,7 @@
 #include <sys/types.h>
 #include <stdio.h>
 #include <freerdp/listener.h>
+#include <winpr/wtypes.h>
 #include <glib-object.h>
 
 typedef struct _DrdServerRuntime DrdServerRuntime;
@@ -17,6 +18,7 @@ G_DECLARE_FINAL_TYPE(DrdRdpSession, drd_rdp_session, DRD, RDP_SESSION, GObject)
 DrdRdpSession *drd_rdp_session_new(freerdp_peer *peer);
 void drd_rdp_session_set_peer_state(DrdRdpSession *self, const gchar *state);
 void drd_rdp_session_set_runtime(DrdRdpSession *self, DrdServerRuntime *runtime);
+void drd_rdp_session_set_virtual_channel_manager(DrdRdpSession *self, HANDLE vcm);
 BOOL drd_rdp_session_post_connect(DrdRdpSession *self);
 BOOL drd_rdp_session_activate(DrdRdpSession *self);
 BOOL drd_rdp_session_pump(DrdRdpSession *self);